Adding an option to get rid of banners in your production files (#39)
* Adding the banner option

The generator ask now if you want to use a banner in your production files. By default it is no. It does not store your last input because in this case it makes no sense to remember your last input AND have a default. That would be confusing.

* Adding the banner into the configs

The banner.config changes its value depending on your inputs by setting up the project

* Adding a test for the banner option, added the store attribut to the banner option.

* Changing some messages for better understanding

* Updated the README

* Changed the position of the banner option

* Removed a duplicated attribute

Removed the duplicated config.banner

* Added the new banner option

See in headline

* Fixing a copy & paste Error

Changed "config" to "props".