feat(baumeister): Move tooling configs to package.json
This way we can get rid of the following files:

```
/.babelrc
/.eslintrc.json
/.prettierrc
/.stylelintrc.json
/baumeister.json
```

It’s a trend that tools make it possible to store their settings in a field in the `package.json`. I recently enabled Baumeister to handle his configuration this way. See [#270 – feat: enable to define baumeister config in package.json](https://github.com/micromata/Baumeister/pull/270).

I’d like the approach, because the project root is less cluttered with dotfiles and you can see all the configs in one place.